day 5, part 2 - doesn't complete
This is how it is implemented by adapting my part 1 solution. It gets
the correct result on the sample data, but doesn't complete on the full
input.

I included a first optimization (switching input list to an iterator)
but it's not enough.

Committing now as it's late and my battery is running out.